Fix infinite recursion while loading MainWindow v2
ApmFirmwareConfig triggered a timer to load the serialPorts
but that should be only populated when the interface is visible,
so I removed the timer call on the constructor.

Also, on the hideEvent of ApmFirmwareConfig we were unconditionally
trying to call MainWindow::instance(), and in some cases the first
call to ::instance() wasn't finished yet ( still on the MainWindow
constructor ) - so we should only act on the hideEvent if m_timer
is active.

Signed-off-by: Tomaz Canabrava <tomaz.canabrava@intel.com>